# docs/api.yaml
openapi: 3.0.1
info:
  title: NAPA project
  description: NAPA trainee assignment
  version: "1"
...
paths:
...
  /ports:
    post:
      tags: ["ports"]
      description: Add a new port to the data base
      operationId: addPort
      requestBody:
       description: Port details
       required: true
       content:
        application/json:
          schema:
            type: object
            description: The name and country of the new port
            properties:
              name:
                $ref: '#/components/schemas/PortName'
              country:
                $ref: '#/components/schemas/CountryName'
      responses:
        '201':
          description: Port created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortName'
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'  
    get:
      tags: ["ports"]
      description: Get the names of all the ports in the data base
      operationId: getPorts
      responses:
        '200':
          description: The list of all port names.
          content:
            application/json:
              schema:
                description: Country names array
                type: array
                items:
                  $ref: '#/components/schemas/PortName'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'  
  /ports/{name}:
    parameters:
     - name: name
       in: path
       required: true  
       description: The name for a port.
       schema:
        $ref: '#/components/schemas/PortName'
    delete:
      tags: ["ports"]
      description: Delete a port
      operationId: deletePort
      responses:
        '204':
          description: Successful delete
        '404':
          description: Port not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'  
    put:
      tags: ["ports"]
      description: Updates a port's name
      operationId: updatePortName
      requestBody:
       description: New port name
       required: true
       content:
        application/json:
          schema:
           $ref: '#/components/schemas/PortName'
      responses:
        '204':
          description: Successful update
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
         description: Forbidden. Port name already exists
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Port not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'  
  /ports/{name}/country:
    parameters:
     - name: name
       in: path
       required: true  
       description: The name for a port.
       schema:
        $ref: '#/components/schemas/PortName'
    put:
      tags: ["ports"]
      description: Update a port's country ?!
      operationId: updatePortCountry
      requestBody:
       description: New country name
       required: true
       content:
        application/json:
          schema:
           $ref: '#/components/schemas/CountryName'
      responses:
        '204':
          description: Successful update
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Port not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'  
    get:
      tags: ["ports"]
      description: Get a port's country
      operationId: getPortCountry
      responses:
       '200':
         description: The name of the country
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/CountryName'
       '404':
          description: Port not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
       '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
components: 
  schemas:
...
    CountryName:
      type: string
      description: The name of a country. Also functions as unique identifier.
      pattern: '^.*?$'
      example: France
...
    PortName:
      type: string
      description: The name of a port.  Also functions as unique identifier.
      pattern: '^.*?$'
      example: Brindisi
...
    ErrorMessage:
      type: string
      pattern: '^.*?$'
      example: Error
      description: Error Message